Updates
1. Create Quiz function is now working. Validations not added yet.
    -> Quiz will only be created after the process is filled completely
    -> Adding new questions will save to a session until u complete the creation process. Else, it will be gone if user leave the page

2. Updated DB schema to remove pk in QuizPrerequisite

+++ app/LearnHub/LearnHub/AppCode/dao/Course_elearnDAO.cs
@@ -485,7 +485,7 @@ namespace LearnHub.AppCode.dao
                         toReturn.setPrerequisite(list); //retrieve arraylist of all prereq course_elearn objects
                     }
                     toReturn.setCategoryID((int)dr["categoryID"]);//7
-                    toReturn.setHoursAwarded((int)dr["hoursAwarded"]);
+                    toReturn.setHoursAwarded((double)dr["hoursAwarded"]);
                     if (!dr.IsDBNull(11))
                     {
                         toReturn.setTargetAudience((string)dr["targetAudience"]);
@@ -612,7 +612,7 @@ namespace LearnHub.AppCode.dao
                         toReturn.setPrerequisite(list); //retrieve arraylist of all prereq course_elearn objects
                     }
                     toReturn.setCategoryID((int)dr["categoryID"]);//7
-                    toReturn.setHoursAwarded((int)dr["hoursAwarded"]);
+                    toReturn.setHoursAwarded((double)dr["hoursAwarded"]);
                     if (!dr.IsDBNull(11))
                     {
                         toReturn.setTargetAudience((string)dr["targetAudience"]);
